Move tests to mirror new location of corresponding source files
